+++ GameService/src/main/java/br/ufg/ceia/gameinsight/gameservice/domain/company/Company.java
@@ -34,11 +34,11 @@ public class Company implements Serializable {
      */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private Instant updatedAt;
 
-    private Long igdbId;
+    private Integer igdbId;
 
     /**
      * The company logo URL.
@@ -75,7 +75,7 @@ public class Company implements Serializable {
      * @param name The name of the company.
      * @param description The description of the company.
      */
-    public Company(Long id, String logoUrl, String name, String description) {
+    public Company(Integer id, String logoUrl, String name, String description) {
         this.id = id;
         this.logoUrl = logoUrl;
         this.name = name;
@@ -98,7 +98,7 @@ public class Company implements Serializable {
      * Get the unique identifier of the company.
      * @return The unique identifier of the company.
      */
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -106,7 +106,7 @@ public class Company implements Serializable {
      * Set the unique identifier of the company.
      * @param id The unique identifier of the company.
      */
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -243,11 +243,11 @@ public class Company implements Serializable {
     }
 
 
-    public Long getIgdbId() {
+    public Integer getIgdbId() {
         return igdbId;
     }
 
-    public void setIgdbId(Long igdbId) {
+    public void setIgdbId(Integer igdbId) {
         this.igdbId = igdbId;
     }
